+++ src/Marketplace.php
@@ -81,6 +81,8 @@ class Marketplace extends BasePlugin
             __METHOD__
         );
         
+        $this->_reviseOrderTemplate();
+        
         // Register provider
         Event::on(
             Providers::class,
@@ -402,6 +404,41 @@ class Marketplace extends BasePlugin
                 }
             }
         );
+    }
+    
+    /* Adds a quick demo of using template hooks to add the
+     * payee to the order editing page. Could use this to try
+     * and modify permissions (ex. redirect you away if you aren’t
+     * supposed to be able to view this order), but might also might
+     * still work better to add a whole new Marketplace tab that shows you
+     * the same default order views as Commerce, but filtered down to only 
+     * show your orders. */
+    private function _reviseOrderTemplate()
+    {
+      Craft::$app->getView()->hook('cp.commerce.order.edit.main-pane', function(array &$context) {
+
+        // This is a demo using the same logic already in
+        // MarketplaceConnectButton_input. Would want to turn this
+        // into Twig if we actually go ahead with using it.
+        $payeeHandle = Marketplace::$plugin->handlesService->getPayeeHandle();
+        $order = $context['order'];
+        if ($order && $order['lineItems'] && sizeof($order['lineItems']) >= 1) {
+          $firstLineItem = $order['lineItems'][0];
+          $product = $firstLineItem['purchasable']['product'];
+          if ($product) {
+            $payeeId = $product[$payeeHandle];
+            if ($payeeId) {
+              $payee = \craft\elements\User::find()
+                ->id($payeeId)
+                ->one();
+              
+              if ($payee) {
+                return '<strong>Payee</strong> ' . $payee;                
+              }
+            }
+          }
+        }
+      });
     }
     
     /**